requirements.txt
# MIT License - ComicAI Dependencies
# Core Framework
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3

# AI and Machine Learning
requests==2.31.0
httpx==0.25.2
Pillow==10.1.0

# Local LLM Support
ollama==0.1.7

# Hugging Face Integration
transformers==4.35.2
torch==2.1.1
diffusers==0.24.0

# Image Processing
opencv-python==4.8.1.78
matplotlib==3.8.2

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
typing-extensions==4.8.0

# Development and Testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: Voice Input Support (install manually if needed)
# SpeechRecognition==3.10.0
# pyaudio==0.2.11  # Install manually: brew install portaudio && pip install pyaudio 